Struct taglib::AudioProperties

pub struct AudioProperties {
    // some fields omitted
}

Common audio file properties.

Instances of AudioProperties can only be created through the taglib::File::audioproperties() method.

Methods

impl AudioProperties

fn length(&self) -> u32

Returns the length, in seconds, of the track.

fn bitrate(&self) -> u32

Returns the most appropriate bit rate for the track, in kB/s. For constant bit rate formats, the returned value is the bit rate of the file; for variable bit rate formats this is either the average or the nominal bit rate.

fn samplerate(&self) -> u32

Returns the sample rate, in Hz.

fn channels(&self) -> u32

Returns the number of audio channels.