Struct taglib::Tag

pub struct Tag<'a> {
    // some fields omitted
}

The abstract meta-data container for audio files

Each Tag instance can only be created by the taglib::File::tag() method.

Methods

impl<'a> Tag<'a>

fn title(&self) -> String

Returns the track name, or an empty string if no track name is present.

fn set_title(&mut self, title: &str)

Sets the track name.

fn artist(&self) -> String

Returns the artist name, or an empty string if no artist name is present.

fn set_artist(&mut self, artist: &str)

Sets the artist name.

fn album(&self) -> String

Returns the album name, or an empty string if no album name is present.

fn set_album(&mut self, album: &str)

Sets the album name.

fn comment(&self) -> String

Returns the track comment, or an empty string if no track comment is present.

fn set_comment(&mut self, comment: &str)

Sets the track comment.

fn genre(&self) -> String

Returns the genre name, or an empty string if no genre name is present.

fn set_genre(&mut self, genre: &str)

Sets the genre name.

fn year(&self) -> u32

Returns the year, or 0 if no year is present.

fn set_year(&mut self, year: u32)

Sets the year.

fn track(&self) -> u32

Returns the track number, or 0 if no track number is present.

fn set_track(&mut self, track: u32)

Sets the track number.